requirement.txt
aniso8601==9.0.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-RESTful==0.3.9
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
pytz==2022.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
python-dotenv==0.20.0
requests==2.27.1